sha-crypt: extract `simple` module
Moves all functionality in the feature-gated `simple` API into its own
module, which eliminates the need for a lot of gating, and means the
"pure" functions sit in lib.rs without additional complications related
to MCF encoding/decoding.
